Add authentication mutation
To be able to generate an initial token that can be send via `Authorization: Bearer <token>`  we need to provide a mutation to create such a token. This is possible by providing username & password or via a refresh token, depending on the use cases and possibilities of the connector.

In addition the `CustomerAddress` was wrongly named `Address` which resulted in a duplicated definition for addresses. Maybe we can set up a Github action to validate the graphql schema to avoid such issues in the future (e.g.: https://github.com/mandiwise/graphql-operation-validation-action)